% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memo_val_percent.R
\name{memo_val_percent}
\alias{memo_val_percent}
\title{Calculate a value for inclusion in a memo, formatted as a percentage}
\usage{
memo_val_percent(
  question,
  values,
  dataset = responses,
  wt = weights_final_trimmed,
  ...
)
}
\arguments{
\item{question}{A survey question name}

\item{values}{Response option value(s) to report. These are usually numeric and should match the questionnaire.}

\item{dataset}{A dataset of survey responses}

\item{wt}{A numeric variable of case weights}

\item{...}{Additional arguments to pass to underlying freqs call}
}
\value{
The weighted percentage of survey respondents that selected the specified values of the target question
}
\description{
Calculate a value for inclusion in a memo, formatted as a percentage
}
\examples{

responses <- data.frame(
s_vote_choice = c(1, 2, 2, 3, 4, 2, NA),
s_vote_likelihood = c(1, 2, 2, 3, 4, 1, NA),
weights_final_trimmed = c(0.9, 0.9, 1.1, 1.1, 1, 1, 1)
)

memo_val_percent(question = s_vote_choice, values = 1)
memo_val_percent(question = s_vote_choice, values = c(1:2))

# Unweighted frequencies are possible but generally not recommended for memos
memo_val_percent(question = s_vote_choice, values = 1, wt = NULL)

}
