% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memo_val_points.R
\name{memo_val_points}
\alias{memo_val_points}
\title{Calculate a value for inclusion in a memo, formatted as a whole number of percentage points}
\usage{
memo_val_points(
  question,
  values,
  dataset = responses,
  wt = weights_final_trimmed,
  group_var = NULL,
  group_values = NULL,
  ...
)
}
\arguments{
\item{question}{A survey question name}

\item{values}{Response option value(s) to report. These are usually numeric and should match the questionnaire.}

\item{dataset}{A dataset of survey responses}

\item{wt}{A numeric variable of case weights}

\item{group_var}{Variable that specifies a subgroup}

\item{group_values}{Subgroup value(s)}

\item{...}{Additional arguments to pass to underlying freqs call}
}
\value{
The weighted percentage points of survey respondents that selected the specified values of the target question
}
\description{
Calculate a value for inclusion in a memo, formatted as a whole number of percentage points
}
\examples{


responses <- data.frame(
s_vote_choice = c(1, 2, 2, 3, 4, 2, NA),
s_vote_likelihood = c(1, 2, 2, 3, 4, 1, NA),
weights_final_trimmed = c(0.9, 0.9, 1.1, 1.1, 1, 1, 1)
)

memo_val_points(question = s_vote_choice, values = 1)
memo_val_points(question = s_vote_choice, values = c(1:2))

# Unweighted frequencies are possible but generally not recommended for memos
memo_val_points(question = s_vote_choice, values = 1, wt = NULL)

}
